Updated precommit and changed dependencies
