+++ src/popup/hooks/useActiveTabTime.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import { get7DaysPriorDate, getIsoDate } from '@shared/utils/dates-helper';
+import { get7DaysPriorDate, getIsoDate } from '@shared/utils/date';
 
 import { usePopupContext } from './PopupContext';
