+++ src/popup/pages/activity/ActivityDatePicker.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { Button, ButtonType } from '@shared/blocks/Button';
 import { Icon, IconType } from '@shared/blocks/Icon';
-import { getDaysInMs, getIsoDate } from '@shared/utils/dates-helper';
+import { getDaysInMs, getIsoDate } from '@shared/utils/date';
 
 interface ActivityDatePickerProps {
   date: string;
