+++ src/shared/services/browser-api/errors.ts
