+++ src/shared/services/browser-api/types.ts
@@ -1,0 +1,3 @@
+export type TabActiveInfo = chrome.tabs.TabActiveInfo;
+export type Tab = chrome.tabs.Tab;
+export type IdleState = chrome.idle.IdleState;
