+++ src/shared/services/browser-api/windows.ts
